+++ src/helpers/OTPublisherHelper.js
@@ -33,6 +33,19 @@ const sanitizeVideoSource = (videoSource = 'camera') => (videoSource === 'camera
 const sanitizeAudioBitrate = (audioBitrate = 40000) =>
   (audioBitrate < 6000 || audioBitrate > 510000 ? 40000 : audioBitrate);
 
+const sanitizeVideoContentHint = (sanitizeVideoContentHint = '') => {
+  switch (sanitizeVideoContentHint) {
+    case 'motion':
+      return 'motion';
+    case 'detail':
+      return 'detail';
+    case 'text':
+      return 'text';
+    default:
+      return '';
+  }
+};
+
 const sanitizeProperties = (properties) => {
   if (typeof properties !== 'object') {
     return {
@@ -46,8 +59,10 @@ const sanitizeProperties = (properties) => {
       cameraPosition: 'front',
       audioFallbackEnabled: true,
       audioBitrate: 40000,
+      enableDtx: false,
       frameRate: 30,
       resolution: sanitizeResolution(),
+      videoContentHint: '',
       videoSource: 'camera',
     };
   }
@@ -56,14 +71,16 @@ const sanitizeProperties = (properties) => {
     audioTrack: sanitizeBooleanProperty(properties.audioTrack),
     publishAudio: sanitizeBooleanProperty(properties.publishAudio),
     publishVideo: sanitizeBooleanProperty(properties.publishVideo),
-    backgroundBlur: sanitizeBooleanProperty(properties.backgroundBlur),
-    pixelatedFace: sanitizeBooleanProperty(properties.pixelatedFace),
+    backgroundBlur: sanitizeBooleanProperty(properties.backgroundBlur, false),
+    pixelatedFace: sanitizeBooleanProperty(properties.pixelatedFace, false),
     name: properties.name ? properties.name : '',
     cameraPosition: sanitizeCameraPosition(properties.cameraPosition),
     audioFallbackEnabled: sanitizeBooleanProperty(properties.audioFallbackEnabled),
     audioBitrate: sanitizeAudioBitrate(properties.audioBitrate),
+    enableDtx: sanitizeBooleanProperty(properties.enableDtx ? properties.enableDtx : false),
     frameRate: sanitizeFrameRate(properties.frameRate),
     resolution: sanitizeResolution(properties.resolution),
+    videoContentHint: sanitizeVideoContentHint(properties.videoContentHint),
     videoSource: sanitizeVideoSource(properties.videoSource),
   };
 };