+++ CHANGELOG.md
@@ -1,3 +1,71 @@
+# 2.25.2 (July 5 2023)
+
+- [Fix]: Fix crash on iOS when publishing a screen-sharing stream.
+
+# 2.25.1  (June 27 2023)
+
+- [Fix]: Fix camera lifecycle on Android. - issue #645
+
+# 2.25.0  (May 17 2023)
+
+- [Update]: Update OpenTok Android SDK and OpenTok iOS SDK to version 2.25.1.
+
+  Note that with this version, we are pinning the major and minor release versions
+  (2.25) to match the corresponding versions in the OpenTok Android and iOS SDKs.
+
+  For iOS, note that this version supports iOS 13+, removes support for FAT binaries
+  and drops 32-bit support. The OpenTok iOS SDK is now available as the OTXCFramework
+  Pod file. (The OpenTok pod file was for FAT binaries.)
+
+  See the release notes for the OpenTok [ioS SDK](https://tokbox.com/developer/sdks/ios/release-notes.html)
+  and the [Android SDK](https://tokbox.com/developer/sdks/android/release-notes.html).
+
+- [Fix]: Fixes an issue in which applications could not connect to a session when
+  the `proxyUrl` option for OTSession was set. - issue #645
+
+# 0.21.4 (April 12 2023)
+
+- [Update]: Revert OpenTok iOS SDK back 2.23.1. There are issues with
+  linked libraries in the OpenTok iOS SDK v2.24.0+ that cause
+  issues when used in React Native. We are working on a bug fix.
+
+# 0.21.3 (March 2023)
+
+- [Update]: iOS SDK to 2.24.2 and Android to 2.24.2 - issue #629
+
+# 0.21.2 (February 14, 2023)
+
+- [Update]: iOS SDK to 2.23.1 and Android to 2.23.1
+- [Fix]: Fix video freeze issue in Android 13 simulator (API level 33) - issue #628
+
+# 0.21.1 (Oct 14, 2022)
+
+- [Update]: iOS SDK to 2.23.1 and Android to 2.23.1
+- [Update]: min target for iOS is now 12.0
+
+# 0.21.0 (June 14, 2022)
+
+- [Update]: Add DTX Option for Publisher
+- [Update]: Android Native SDK to 2.22.3 and iOS to 2.22.3
+
+# 0.20.3 (May 18, 2022)
+
+- [Fix]: Updates from DependatBot
+
+# 0.20.2 (May 16, 2022)
+
+- [Update]: Readme file update with Bintray instructions
+- [Update]: Android Native SDK to 2.21.5 and iOS to 2.21.3
+- [Fix]: Typescrit types fixed audioLevel
+
+# 0.20.1 (Jan 28, 2022)
+- [Update]: Add Content Hint API on the OTPublisher object
+- [Update]: Add timestamp property on the audioNetworkStats and videoNetworkStats
+- [Fix]: Bump `opentok-react-native` version to 
+
+# 0.20.0 (Jan 26, 2022)
+- [Update]: update SDKs iOS to 2.21.3 and Android to 2.21.4 
+
 # 0.19.0 (Oct 11, 2021)
 - [Update]: update SDKs Android to 2.20.2
 - [Update]: added IceConfig options in the Session Options object